Windows DLL: Expose static and shared variant as an interface lib
systemc can't build as a single shared library on Windows because the
DLL format can't supported unresolved symbols (sc_main).
There the library is split in two part:
- static library that contains unresolved symbols
- dynamic library that contains resolved symbols

The libraries listed below are expored as a CMake target name systemc
to keep the same behaviour as when system is build a single static library.